Modified code clean_data.py
Added variables to drop in clean_data.py. The portion of the code that was responsible for dealing with NAs was separated out as na_fix.py. Additionally the engineering part was also separated out as engineer.py, and was changed quite a bit. Additionally added location.py that calculates the distance to the nearest metro and the city center. The location based code will be expanded soon.