upgrading duplicate code
Implemented a new code for removing duplicates. This new code relies on another method to double check when the first method claims that the house is not a duplicate.

Additionally, the code now fetches images as it checks them. Instead of first fetching all the images and then comparing them between the observations.

Additionally the comparison has become stricter to better be able to detect duplicates.